# api/openapi.yaml
openapi: 3.0.3
info:
  title: Geek Budget - OpenAPI 3.0
  contact:
    email: ilya.korolev@outlook.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
paths:
...
  /v1/accounts/{id}:
    get:
      tags:
        - accounts
      summary: get account
      operationId: getAccount
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the account"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    put:
      tags:
        - accounts
      summary: update account
      operationId: updateAccount
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the account"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountNoID"
      responses:
        "200":
          description: updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      tags:
        - accounts
      summary: delete account
      operationId: deleteAccount
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the account"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: no body
...
  /v1/transactions/{id}:
    put:
      tags:
        - transactions
      summary: update transaction
      operationId: updateTransaction
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionNoID"
      responses:
        "200":
          description: updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    get:
      tags:
        - transactions
      summary: get transaction
      operationId: getTransaction
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      tags:
        - transactions
      summary: delete transaction
      operationId: deleteTransaction
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: no body
...
  /v1/matchers:
    post:
      tags:
        - matchers
      summary: create new matcher
      operationId: createMatcher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatcherNoID"
      responses:
        "200":
          description: created matcher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matcher"
    get:
      tags:
        - matchers
      summary: get all matchers
      operationId: getMatchers
      responses:
        "200":
          description: matchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Matcher"
  /v1/matchers/{id}:
    get:
      tags:
        - matchers
      summary: get matcher
      operationId: getMatcher
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the matcher"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: matcher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matcher"
    put:
      tags:
        - matchers
      summary: update matcher
      operationId: updateMatcher
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the matcher"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatcherNoID"
      responses:
        "200":
          description: updated matcher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matcher"
    delete:
      tags:
        - matchers
      summary: delete matcher
      operationId: deleteMatcher
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the matcher"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: no body
...
  /v1/matchers/check:
    post:
      tags:
        - matchers
      summary: check if passed matcher matches given transaction
      operationId: checkMatcher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                matcher:
                  $ref: "#/components/schemas/MatcherNoID"
                transaction:
                  $ref: "#/components/schemas/TransactionNoID"
              required:
                - matcher
                - transaction
      responses:
        "200":
          description: true/false if matcher works
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                  reason:
                    type: string
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Entity:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      required:
        - id
...
    BankAccountInfo:
      type: object
      properties:
        accountId:
          type: string
          example: "12345"
        bankId:
          type: string
          example: "12345"
        balances:
          description: >-
            List of balances for this account. It's an array since one account could hold multiple
            currencies, for example, cash account could hold EUR, USD and CZK. Or one bank account could
            hold multiple currencies.
          type: array
          items:
            type: object
            properties:
              currencyId:
                type: string
                format: uuid
              openingBalance:
                type: number
                format: double
                example: 123.0
              closingBalance:
                type: number
                format: double
                example: 321.0
    AccountNoID:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          enum:
            - expense
            - income
            - asset
        bankInfo:
          $ref: "#/components/schemas/BankAccountInfo"
        showInDashboardSummary:
          type: boolean
          description: "If true, show this account in dashboard summary."
          default: true
        image:
          type: string
          description: "ID of the account image"
      required:
        - name
        - type
    Account:
      type: object
      allOf:
        - $ref: "#/components/schemas/Entity"
        - $ref: "#/components/schemas/AccountNoID"
      required:
        - showInDashboardSummary
...
    TransactionNoID:
      type: object
      properties:
        date:
          type: string
          format: date-time
        description:
          type: string
        place:
          type: string
        tags:
          type: array
          items:
            type: string
        partnerName:
          type: string
        partnerAccount:
          type: string
        partnerInternalId:
          type: string
          description: Internal bank's ID to be able to match later if necessary
        extra:
          type: string
          description: >-
            Stores extra data about transaction. For example could hold "variable symbol" to
            distinguish payment for the same account, but with different meaning
        unprocessedSources:
          type: string
          description: >-
            Stores FULL unprocessed transactions which was source of this transaction.
            Could be used later for detailed analysis
        externalIds:
          type: array
          items:
            type: string
          description: >-
            IDs of unprocessed transaction - to match later
        movements:
          type: array
          items:
            $ref: "#/components/schemas/Movement"
        matcherId:
          type: string
          format: uuid
          description: ID of the matcher used for this conversion (if any)
        isAuto:
          type: boolean
          description: If true, this transaction was converted automatically by the matcher
      required:
        - date
        - movements
    Transaction:
      type: object
      allOf:
        - $ref: "#/components/schemas/Entity"
        - $ref: "#/components/schemas/TransactionNoID"
    Movement:
      type: object
      properties:
        amount:
          type: number
          format: double
        currencyId:
          type: string
        accountId:
          type: string
        description:
          type: string
      required:
        - amount
        - currencyId
        # accountId is not required to create unprocessed transactions via API
        # - accountId
...
    MatcherNoID:
      type: object
      properties:
        outputDescription:
          type: string
        outputAccountId:
          type: string
          format: uuid
        outputTags:
          type: array
          items:
            type: string

        currencyRegExp:
          type: string
        partnerNameRegExp:
          type: string
        partnerAccountNumberRegExp:
          type: string
        descriptionRegExp:
          type: string
        extraRegExp:
          type: string
        confirmationHistory:
          type: array
          items:
            type: boolean
          description: >-
            List of booleans representing manual confirmations for this matcher
            (true = confirmed, false = rejected). Server enforces maximum length
            configured via application config.
        image:
          type: string
          description: "ID of the matcher image"
      required:
        - outputDescription
        - outputAccountId
    Matcher:
      type: object
      allOf:
        - $ref: "#/components/schemas/Entity"
        - $ref: "#/components/schemas/MatcherNoID"
        - type: object
          properties:
            confirmationsCount:
              type: integer
              description: >-
                Number of successful confirmations (true values) in the confirmation history.
                This shows how many times the matcher was confirmed as correct.
              example: 9
            confirmationsTotal:
              type: integer
              description: >-
                Total length of the confirmation history array.
                This is the total number of times this matcher has been evaluated.
              example: 10
          required:
            - confirmationsCount
            - confirmationsTotal
...